Fix dtache initialization
Fix dtache-initialize. The session directory needs to always be
checked and maybe created. The reason is that the user can have
initialized the connection to the database whilst on a remote
machine. Then on the local machine the check for initialization of the
database will be skipped, this patch makes sure that doesn't happen.