Merge develop branch into master
This commit contains a lot of recent improvements. To mention a few of
them:

- Dtache now has proper support for running on remote hosts
- Neither dtache.el nor dtache-shell.el have any external dependencies
- A command to list the sessions using a tabulated list interface now
if one doesn't want to use the completing-read interface
- Dtache now requires at least Emacs 27.1. This is to allow proper
customization for remote sessions
- Sessions are now launched asynchronously using start-file-process,
which takes the current default-directory into account
- Notification functionality has been implemented to notify users when
a session finishes
- dtache.el now exposes two different functions for usage of
others. The first one is dtache-shell-command, which is similar to
async-shell-command and is supposed to be used by the user
directly. The other one is dtache-start-process which targets usage
from other packages
- Sessions have been improved so that they now can store closures for
custom open/callback functionality
- The usage of the sql database has been dropped and instead reworked
to save the lisp objects to file directly for persistent storage
- Annotators have been reworked