Add package dtache-eshell
This patch adds the dtache-eshell package which is an integration of
dtache into eshell. This package is similar in functionality to the
dtache-shell package.

The only difference from a users perspective is that the user doesn't
need to use a special binding to detach from a session. C-c C-c will
do that for you.