Improve dtache-shell-command
This patch updates the internals of the dtache-shell-command. The new
approach is to utilize the async-shell-command and let dtache create
the altered command to run.

The dtache-shell-mode is moved from dtache-shell into dtache which
makes it possible to enable the mode in dtache-initialize. This is
necessary because async-shell-command is also run in shell-mode so the
mode is utilized there as well, which makes it possible for users to
reuse their detach binding that they use in M-x shell