Improve error handling with dtache-valid-session
Make commands use dtache-valid-session to make sure that the session
provided to the function actually is a session object, and if so that
it can be found. If it is not found it is indicative that the host
that the missing session belongs to is not up to date. So we trigger
an automatic cleanup for that session.