Rename dtache-initialize to dtache-setup
All extensions have dtache-.*-setup functions that should be used. But
the core package had dtache-initialize. This patch renames the core
function in order to harmonize the naming.

It also makes sure the other setup functions runs dtache-setup, which
in case of that has already been called will do nothing.