Add version to the dtache database
This patch adds a header to the dtache.db file. This header allows
dtache to read the dtache-session version of the sessions stored in
the database.

If the version is not matching the databasee content is
discarded. The version should be increase when an update has been made
to the dtache-session object, which breaks backwards compatibility.