Monitor updates to the database
This patch utilizes filenotify to add a watch for the
dtache-db-directory. This is needed in order for dtache to support
multiple active Emacs sessions. Any change to the database by any of
the Emacs instances will trigger a re-initialization of all
dtache-sessions. This means that once a session is created from any of
the Emacs sessions, that information will propagate to the others as well.