Disable dtach from redrawing
Normally, dtach will set the redraw flag to `none`, because dtache
doesn't run programs in a terminal. If dtache-env is edited to use the
`script` command, though, dtach will redraw using the ctrl_l
method. Emacs doesn't handle this character well, so we should disable
it for a more uniform experience