Improve dtache integration with eshell
This commit changes the integration towards eshell. Instead of relying
on a hook we are using and advice. With the advice we can target the
eshell-external-command function, which should be called with data
that is supposed to be run by an external shell. By integrating on
this level we make sure that eshell aliases have been expanded.