Bug fix: Make advice preserve called-interactively
This patch addresses a subtle bug that was present in
detached-shell-override-history.

The issue was that apply was used to call the original function in the
advice. This however lead to orig-fun, which is shell, not detect that
it was called interactively.

The called-interactively-p is used in shell, and the docstring of the
function says:

"This function is very brittle, it may fail to return the intended result when
the code is debugged, advised, or instrumented in some form."

Which explains why it failed. The issue became noticeable on remote
hosts where shell would not ask the user for a "Remote shell path "
since it didn't think it was called interactively.