Bug fix: Properly remove notify watches
This patch addresses a bug in the package which was caused by
alist-get failing to retrieve the watch due to key being a string. The
test function needs to be explicitly set in order for it to work. This
caused detached to keep on adding new watches for the session
directory resulting in multiple notifications when a session finished.