Improve handling of remote sessions
This patch updates the initialization of remote sessions. The package
will now only watch sessions that are local, or currently
accessible. This should mean that if detached is loaded directly when
Emacs starts it won't try to access remote sessions. Instead it will
wait until the user either switch to the remote host, or access the
session willingly through detached-open-session.