Change default value for session/db directory
Using temporary-file-directory to store sessions has had the neat
property that on reboot the old sessions will be automatically
wiped. And users have the possibility to configure this location
themselves.

However there are security concerns of using the
temporary-file-directory if other users have access to the same
computer. So instead we change to user-emacs-directory and make sure
that sessions as well as the database is written to a sub folder named
detached.