Speed up initialization of sessions
Change detached--initialize-session to not update the database each
time its called. When the user has many sessions this creates a lot of
overhead.

The new approach is instead to add a call to
detached--db-update-sessions after all the individual sessions has
been initialized which should then only be called once per session
directory initialization.