Make detached-list apply first filter if defined
If the user has customized the detached-list-filters, the first filter
in that list will used as the default filter when detached-list is
first called. This allows the user to define a default filter. If the
user wants to have multiple filters but want all sessions to be
visible by default they would simply add the following filter first in
the list.

("All" . nil)