New implementation for killing a session
The previous implementation relied on finding all the child processes
to the dtach process and terminating them in reverse order. This
approach worked most of the time, but in situations where the initial
process had forked of such as would happen when a docker container
started detached failed to terminate the session properly.

The new approach instead relies on the fact that when we are attached
to a session in a comint derived mode we can simply execute
comint-interrupt-subjob. When not attached to the session we instead
attach to it in the background by creating a dedicated comint buffer
and attaching to the session and sending comint-interrupt-subjob. For
this to work reliably we add a small delay between the steps using
timers.